Move items into atomic design dir structure
