+++ src/components/atoms/LogoChroma.js
