+++ src/components/atoms/ShadowBox.stories.js
@@ -3,4 +3,4 @@ import { storiesOf } from '@storybook/react';
 import { Button } from '@storybook/design-system';
 import ShadowBox from './ShadowBox';
 
-storiesOf('ShadowBox', module).add('base', () => <ShadowBox>Shadow box content</ShadowBox>);
+storiesOf('ShadowBox', module).add('default', () => <ShadowBox>Shadow box content</ShadowBox>);
