+++ src/components/organisms/Footer.stories.js
@@ -1,0 +1,5 @@
+import React from 'react';
+import { storiesOf } from '@storybook/react';
+import Footer from './Footer';
+
+storiesOf('Footer', module).add('default', () => <Footer />);
